Moving things around for directory structure. Updating CL and PJ for changes
