+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "events-admin-ui",
-  "version": "0.7.0",
+  "version": "0.8.0",
   "private": true,
   "scripts": {
     "dev": "next dev",
@@ -16,8 +16,7 @@
     "react-dom": "^18",
     "react-hook-form": "^7.53.1",
     "react-redux": "^9.1.2",
-    "zod": "^3.23.8",
-    "zustand": "^5.0.0"
+    "zod": "^3.23.8"
   },
   "devDependencies": {
     "@types/node": "^20",
