+++ src/lib/hooks.ts
@@ -1,6 +1,7 @@
 // This file serves as a central hub for re-exporting pre-typed Redux hooks.
 import { useDispatch, useSelector, useStore } from "react-redux";
-import type { AppDispatch, AppStore, RootState } from "./store";
+
+import type { AppDispatch, AppStore, RootState } from "@/lib/store";
 
 // Use throughout your app instead of plain `useDispatch` and `useSelector`
 export const useAppDispatch = useDispatch.withTypes<AppDispatch>();
